test: split test_tools_kitchen_visibility.py into 4 per-concern files
Split the 872-line test_tools_kitchen_visibility.py by concern
(Group C integration / Group H sous-chef / Group I redisable unit /
T-VIS-003/004/005 feature gates) into:

- test_tools_kitchen_visibility.py (kept, trimmed ~411 LoC): T-VISIBILITY-1a/1b/2/2b/3a/3
  Group C integration tests + 3 Group I _redisable_subsets unit tests
- test_tools_kitchen_recipe_admission.py (~244 LoC): 7 functions covering the
  open_kitchen(name=...) admission control flow
- test_tools_kitchen_feature_gates.py (~109 LoC): T-VIS-003/004/005 subset
  re-disable integration tests
- test_tools_kitchen_sous_chef.py (~158 LoC): 4 Group H sous-chef discipline
  injection tests

Mechanical lift: no production code changes, no behavior changes, no test
assertion changes. Each file carries pytestmark = [layer("server"), small].
Group markers (Group C / H / I) redistributed into per-concern files.

Refs: #4610
