+++ tests/server/test_tools_kitchen_visibility.py
@@ -1,4 +1,12 @@
-"""Tests for tools_kitchen.py: visibility, component management, sous-chef, redisable_subsets."""
+"""Tests for tools_kitchen visibility and component management (kept file).
+
+Carries the Group C integration tests (open_kitchen / close_kitchen backend
+visibility, client-cache coherence roundtrips, gate-file non-production) and
+the Group I `_redisable_subsets` unit tests. The recipe-admission, sous-chef
+discipline, and T-VIS-003/004/005 feature-gate integration tests live in their
+own per-concern files (`test_tools_kitchen_recipe_admission.py`,
+`test_tools_kitchen_feature_gates.py`, `test_tools_kitchen_sous_chef.py`).
+"""
 
 from __future__ import annotations
 
@@ -7,21 +15,10 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from autoskillit.core import KITCHEN_GATED_TOOLS, RecipeNotFoundError
-from autoskillit.core.types._type_constants import SOUS_CHEF_MANDATORY_SECTIONS
-from autoskillit.pipeline import (
-    bind_kitchen_intent,
-    claim_kitchen_request,
-    release_kitchen_request,
-)
-from tests.server._helpers import _configure_admitted_recipe, _with_finalized_projection
+from autoskillit.core import KITCHEN_GATED_TOOLS
 from tests.server.conftest import _make_mock_ctx
 
 pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
-
-# ---------------------------------------------------------------------------
-# Group C — visibility + component management
-# ---------------------------------------------------------------------------
 
 
 # T-VISIBILITY-1a: open_kitchen calls enable_components for notification-capable backend
@@ -344,464 +341,6 @@ async def test_open_kitchen_includes_categorized_tool_listing(tmp_path, monkeypa
                     f"Tool '{tool_name}' missing from open_kitchen response"
                 )
                 seen.add(tool_name)
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_with_recipe_returns_combined_response(tmp_path, monkeypatch):
-    """open_kitchen(name='x') opens kitchen AND loads the recipe in one call."""
-    monkeypatch.chdir(tmp_path)
-    recipes_dir = tmp_path / ".autoskillit" / "recipes"
-    recipes_dir.mkdir(parents=True)
-    yaml_content = (
-        "name: test-recipe\ndescription: test\nsteps:\n  do:\n    tool: run_cmd\n"
-        "    with:\n      cmd: echo hi\n    on_success: done\n    on_failure: done\n"
-        "  done:\n    action: stop\n    message: Done\n"
-    )
-    (recipes_dir / "test-recipe.yaml").write_text(yaml_content)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.load_and_validate.return_value = {
-        "content": yaml_content,
-        "valid": True,
-        "suggestions": [],
-        "diagram": None,
-    }
-    mock_ctx.recipes.load_and_validate.return_value = _with_finalized_projection(
-        mock_ctx.recipes.load_and_validate.return_value
-    )
-    _configure_admitted_recipe(mock_ctx, recipes_dir / "test-recipe.yaml")
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result_str = await open_kitchen(name="test-recipe", ctx=mock_ctx)
-
-    result = json.loads(result_str)
-    assert result["success"] is True
-    assert result["kitchen"] == "open"
-    assert "version" in result
-    assert "content" in result
-    assert "test-recipe" in result["content"]
-    mock_ctx.gate.enable.assert_called_once()
-    mock_ctx.enable_components.assert_called_once_with(tags={"kitchen"})
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_reports_admitted_recipe_missing_during_load(tmp_path, monkeypatch):
-    """An admitted recipe that disappears during loading fails closed."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    _configure_admitted_recipe(mock_ctx, tmp_path / "nonexistent.yaml")
-    mock_ctx.recipes.load_and_validate.side_effect = RecipeNotFoundError(
-        "No recipe named 'nonexistent' found"
-    )
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result_str = await open_kitchen(name="nonexistent", ctx=mock_ctx)
-
-    result = json.loads(result_str)
-    assert result["success"] is False
-    assert result["stage"] == "load_and_validate"
-    assert result["error"] == "RecipeNotFoundError: No recipe named 'nonexistent' found"
-    assert result["kitchen"] == "failed"
-
-
-@pytest.mark.anyio
-@pytest.mark.parametrize(
-    ("recipe_name", "skill_match", "expected_message"),
-    [
-        ("dry-walkthrough", object(), "current session"),
-        ("missing-recipe", None, "No recipe named 'missing-recipe' found"),
-    ],
-)
-async def test_open_kitchen_rejects_non_recipe_before_operational_mutation(
-    tmp_path,
-    monkeypatch,
-    recipe_name,
-    skill_match,
-    expected_message,
-):
-    """Skill-only and unknown names fail before operational state is mutated."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.find.return_value = None
-    mock_ctx.skill_resolver.resolve_effective.return_value = skill_match
-    mock_ctx.recipe_name = "existing-recipe"
-    previous_execution = mock_ctx.recipe_initialization_state
-    previous_quota_task = mock_ctx.quota_refresh_task
-
-    handler = AsyncMock()
-    with (
-        patch("autoskillit.server._get_ctx", return_value=mock_ctx),
-        patch("autoskillit.server.tools.tools_kitchen._open_kitchen_handler", handler),
-        patch("autoskillit.server.tools.tools_kitchen.clear_recipe_execution") as clear,
-        patch("autoskillit.server.tools.tools_kitchen.create_background_task") as create_task,
-        patch(
-            "autoskillit.server.tools.tools_kitchen.bind_kitchen_intent",
-            wraps=bind_kitchen_intent,
-        ) as bind_request,
-        patch(
-            "autoskillit.server.tools.tools_kitchen.claim_kitchen_request",
-            wraps=claim_kitchen_request,
-        ) as claim_request,
-        patch(
-            "autoskillit.server.tools.tools_kitchen.release_kitchen_request",
-            wraps=release_kitchen_request,
-        ) as release_request,
-        patch("autoskillit.server.tools.tools_kitchen.mcp") as mock_mcp,
-    ):
-        from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-        result = json.loads(await open_kitchen(name=recipe_name, ctx=mock_ctx))
-
-    assert result["stage"] == "recipe_namespace"
-    assert expected_message in result["user_visible_message"]
-    handler.assert_not_awaited()
-    clear.assert_not_called()
-    create_task.assert_not_called()
-    mock_mcp.enable.assert_not_called()
-    mock_ctx.recipes.load.assert_not_called()
-    mock_ctx.recipes.load_and_validate.assert_not_called()
-    bind_request.assert_called_once()
-    claim_request.assert_called_once()
-    release_request.assert_called_once()
-    assert mock_ctx.recipe_name == "existing-recipe"
-    assert mock_ctx.recipe_initialization_state is previous_execution
-    assert mock_ctx.quota_refresh_task is previous_quota_task
-    assert mock_ctx.kitchen_open_state.phase.value == "request_bound"
-    assert mock_ctx.kitchen_open_state.request_active is False
-    effects = [(effect.name, effect.phase.value) for effect in mock_ctx.kitchen_open_state.effects]
-    assert effects == [("request_binding", "confirmed")]
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_does_not_misclassify_resolver_runtime_error(tmp_path, monkeypatch):
-    """A resolver failure is not reported as missing server configuration."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.find.return_value = None
-    mock_ctx.skill_resolver.resolve_effective.side_effect = RuntimeError("resolver failed")
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-        result = json.loads(await open_kitchen(name="broken-recipe", ctx=mock_ctx))
-
-    assert result["stage"] == "unhandled"
-    assert result["error"] == "RuntimeError: resolver failed"
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_without_recipe_bypasses_namespace_admission(tmp_path, monkeypatch):
-    """open_kitchen(name=None) activates without consulting recipe or skill names."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes.find.side_effect = AssertionError("anonymous activation admitted a recipe")
-    mock_ctx.skill_resolver.resolve_effective.side_effect = AssertionError(
-        "anonymous activation resolved a skill"
-    )
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result = await open_kitchen(ctx=mock_ctx)
-
-    assert isinstance(result, str)
-    parsed = json.loads(result)
-    assert parsed["success"] is True
-    assert parsed["kitchen"] == "open"
-    assert "Kitchen is open" in parsed["content"]
-    assert parsed["ingredients_table"] is None
-    mock_ctx.recipes.find.assert_not_called()
-    mock_ctx.skill_resolver.resolve_effective.assert_not_called()
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_denied_by_gate_when_headless(tmp_path, monkeypatch):
-    monkeypatch.setenv("AUTOSKILLIT_HEADLESS", "1")
-    monkeypatch.chdir(tmp_path)
-    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-    result = json.loads(await open_kitchen())
-    assert result["success"] is False
-    assert result["kitchen"] == "failed"
-    assert "user_visible_message" in result
-    assert len(result["user_visible_message"]) > 0
-    assert result["stage"] == "headless_guard"
-
-
-@pytest.mark.anyio
-async def test_close_kitchen_denied_when_headless(tmp_path, monkeypatch):
-    monkeypatch.setenv("AUTOSKILLIT_HEADLESS", "1")
-    monkeypatch.chdir(tmp_path)
-    from autoskillit.server.tools.tools_kitchen import close_kitchen
-
-    result = json.loads(await close_kitchen())
-    assert result["success"] is False
-    assert result["subtype"] == "headless_error"
-
-
-# T-VIS-003
-@pytest.mark.anyio
-async def test_open_kitchen_redisables_subsets(tmp_path, monkeypatch):
-    """open_kitchen must call ctx.disable_components for each disabled subset."""
-    from autoskillit.config.settings import AutomationConfig, SubsetsConfig
-
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.disable_components = AsyncMock()
-    mock_ctx.config = AutomationConfig(subsets=SubsetsConfig(disabled=["github", "ci"]))
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    await open_kitchen(ctx=mock_ctx)
-
-    disable_calls = mock_ctx.disable_components.call_args_list
-    disabled_tags = [
-        c.kwargs.get("tags") or (c.args[0] if c.args else None) for c in disable_calls
-    ]
-    assert {"github"} in disabled_tags
-    assert {"ci"} in disabled_tags
-
-
-# T-VIS-004
-@pytest.mark.anyio
-async def test_open_kitchen_redisable_order(tmp_path, monkeypatch):
-    """ctx.disable_components must be called after ctx.enable_components (order matters)."""
-    from autoskillit.config.settings import AutomationConfig, SubsetsConfig
-
-    monkeypatch.chdir(tmp_path)
-    call_order = []
-    mock_ctx = _make_mock_ctx()
-
-    async def record_enable(**kwargs):
-        call_order.append(("enable", kwargs))
-
-    async def record_disable(**kwargs):
-        call_order.append(("disable", kwargs))
-
-    mock_ctx.enable_components = record_enable
-    mock_ctx.disable_components = record_disable
-    mock_ctx.config = AutomationConfig(subsets=SubsetsConfig(disabled=["github"]))
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    await open_kitchen(ctx=mock_ctx)
-
-    enable_idx = next(i for i, (op, _) in enumerate(call_order) if op == "enable")
-    disable_idx = next(i for i, (op, _) in enumerate(call_order) if op == "disable")
-    assert enable_idx < disable_idx, "disable_components must be called after enable_components"
-
-
-# T-VIS-005
-@pytest.mark.anyio
-async def test_open_kitchen_keeps_disabled_feature_tags_hidden_when_subsets_empty(
-    tmp_path, monkeypatch
-):
-    """An empty subset list still keeps disabled feature tags hidden."""
-    from autoskillit.config.settings import AutomationConfig, SubsetsConfig
-
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.disable_components = AsyncMock()
-    mock_ctx.config = AutomationConfig(
-        subsets=SubsetsConfig(disabled=[]), features={"fleet": True}
-    )
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    await open_kitchen(ctx=mock_ctx)
-
-    mock_ctx.disable_components.assert_called_once_with(tags={"exploration"})
-
-
-# ---------------------------------------------------------------------------
-# Group H — sous-chef discipline injection
-# ---------------------------------------------------------------------------
-
-
-@pytest.mark.anyio
-async def test_sous_chef_discipline_not_in_open_kitchen_result(tmp_path, monkeypatch):
-    """open_kitchen does not inject sous_chef_discipline — delivery is via system prompt."""
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.load_and_validate.return_value = {
-        "content": "name: implementation\nsteps:\n  do:\n    tool: run_cmd\n",
-        "valid": True,
-        "suggestions": [],
-        "diagram": None,
-    }
-    mock_ctx.recipes.load_and_validate.return_value = _with_finalized_projection(
-        mock_ctx.recipes.load_and_validate.return_value
-    )
-    _configure_admitted_recipe(mock_ctx, tmp_path / "implementation.yaml")
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result_str = await open_kitchen(name="implementation", ctx=mock_ctx)
-
-    result = json.loads(result_str)
-    assert result["success"] is True
-    assert "sous_chef_discipline" not in result, (
-        "sous_chef_discipline must not be injected by open_kitchen; "
-        "food truck delivery is handled by _build_admiral_dispatch_block() "
-        "in fleet/_prompts.py"
-    )
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_result_keys_match_typed_dict(tmp_path, monkeypatch):
-    """All keys in open_kitchen result are declared in OpenKitchenResult."""
-    from autoskillit.recipe._recipe_ingredients import OpenKitchenResult
-
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.recipes = MagicMock()
-    mock_ctx.recipes.load_and_validate.return_value = {
-        "content": "name: implementation\nsteps:\n  do:\n    tool: run_cmd\n",
-        "valid": True,
-        "suggestions": [],
-        "diagram": None,
-    }
-    mock_ctx.recipes.load_and_validate.return_value = _with_finalized_projection(
-        mock_ctx.recipes.load_and_validate.return_value
-    )
-    _configure_admitted_recipe(mock_ctx, tmp_path / "implementation.yaml")
-    mock_ctx.config.migration.suppressed = []
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result_str = await open_kitchen(name="implementation", ctx=mock_ctx)
-
-    result = json.loads(result_str)
-    declared = set(OpenKitchenResult.__annotations__)
-    undeclared = set(result.keys()) - declared
-    assert undeclared == set(), (
-        f"open_kitchen returned keys not declared in OpenKitchenResult: {sorted(undeclared)}. "
-        "Add each to OpenKitchenResult in recipe/_recipe_ingredients.py."
-    )
-    for key in ("success", "kitchen", "version"):
-        assert key in result, (
-            f"open_kitchen result missing always-present key {key!r}. "
-            "OpenKitchenResult declares it but the handler does not populate it."
-        )
-
-
-@pytest.mark.anyio
-async def test_sous_chef_rules_injected_at_open_kitchen(tmp_path, monkeypatch):
-    """Path B (no-name) must inject full sous-chef SKILL.md into response text."""
-    from autoskillit.execution.backends import ClaudeCodeBackend
-    from autoskillit.workspace import SkillsDirectoryProvider
-
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.project_dir = tmp_path
-    mock_ctx.skill_resolver = SkillsDirectoryProvider().resolver
-    mock_ctx.backend = ClaudeCodeBackend()
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result = await open_kitchen(ctx=mock_ctx)
-
-    parsed = json.loads(result)
-    content = parsed["content"]
-    for header in SOUS_CHEF_MANDATORY_SECTIONS:
-        assert header in content, (
-            f"open_kitchen no-name response missing sous-chef section: {header!r}"
-        )
-
-
-@pytest.mark.anyio
-async def test_open_kitchen_degrades_gracefully_without_sous_chef(tmp_path, monkeypatch):
-    """When sous-chef SKILL.md is absent, Path B must return a valid response without raising."""
-    from types import SimpleNamespace
-
-    monkeypatch.chdir(tmp_path)
-    mock_ctx = _make_mock_ctx()
-    mock_ctx.enable_components = AsyncMock()
-    mock_ctx.skill_resolver.list_effective.return_value = SimpleNamespace(skills=())
-    mock_ctx.skill_resolver.resolve_effective.return_value = None
-
-    with patch("autoskillit.server._get_ctx", return_value=mock_ctx):
-        with patch("autoskillit.server.logger"):
-            with patch(
-                "autoskillit.server.tools.tools_kitchen._prime_quota_cache", new=AsyncMock()
-            ):
-                with patch("autoskillit.server.tools.tools_kitchen._write_hook_config"):
-                    from autoskillit.server.tools.tools_kitchen import open_kitchen
-
-                    result = await open_kitchen(ctx=mock_ctx)
-
-    parsed = json.loads(result)
-    assert parsed["success"] is True
 
 
 # ---------------------------------------------------------------------------
